+++ podchat/src/main/java/com/fanap/podchat/call/CallAsyncRequestsManager.java
@@ -13,6 +13,7 @@ import com.fanap.podchat.call.request_model.GetCallHistoryRequest;
 import com.fanap.podchat.call.request_model.GetCallParticipantsRequest;
 import com.fanap.podchat.call.request_model.MuteUnMuteCallParticipantRequest;
 import com.fanap.podchat.call.request_model.RejectCallRequest;
+import com.fanap.podchat.call.request_model.StartOrEndCallRecordRequest;
 import com.fanap.podchat.call.request_model.TerminateCallRequest;
 import com.fanap.podchat.call.result_model.CallCancelResult;
 import com.fanap.podchat.call.result_model.CallCreatedResult;
@@ -30,9 +31,11 @@ import com.fanap.podchat.call.result_model.RemoveFromCallResult;
 import com.fanap.podchat.call.result_model.StartedCallModel;
 import com.fanap.podchat.chat.App;
 import com.fanap.podchat.chat.CoreConfig;
+import com.fanap.podchat.chat.assistant.model.AssistantVo;
 import com.fanap.podchat.mainmodel.AsyncMessage;
 import com.fanap.podchat.mainmodel.ChatMessage;
 import com.fanap.podchat.mainmodel.Invitee;
+import com.fanap.podchat.mainmodel.Participant;
 import com.fanap.podchat.model.ChatResponse;
 import com.fanap.podchat.requestobject.RequestAddParticipants;
 import com.fanap.podchat.requestobject.RemoveParticipantRequest;
@@ -407,6 +410,51 @@ public class CallAsyncRequestsManager {
         message.setType(ChatMessageType.Constants.END_CALL_REQUEST);
         message.setToken(CoreConfig.token);
         message.setSubjectId(request.getCallId());
+        message.setTokenIssuer(CoreConfig.tokenIssuer);
+        message.setUniqueId(uniqueId);
+        message.setTypeCode(Util.isNullOrEmpty(request.getTypeCode()) ? request.getTypeCode() : CoreConfig.typeCode);
+        JsonObject a = (JsonObject) App.getGson().toJsonTree(message);
+        return a.toString();
+
+
+    }
+
+    public static String createStartRecordCall(StartOrEndCallRecordRequest request, String uniqueId)  throws PodChatException {
+
+        AsyncMessage message = new AsyncMessage();
+        message.setType(ChatMessageType.Constants.START_RECORDE_CALL);
+        message.setToken(CoreConfig.token);
+        message.setSubjectId(request.getSubjectId());
+        message.setTokenIssuer(CoreConfig.tokenIssuer);
+        message.setUniqueId(uniqueId);
+        message.setTypeCode(Util.isNullOrEmpty(request.getTypeCode()) ? request.getTypeCode() : CoreConfig.typeCode);
+        JsonObject a = (JsonObject) App.getGson().toJsonTree(message);
+        return a.toString();
+
+
+    }
+
+    public static ChatResponse<Participant> handleStartedRecordCallResponse(ChatMessage chatMessage) {
+
+        ChatResponse<Participant> response = new ChatResponse<>();
+        Participant result = App.getGson().fromJson(chatMessage.getContent(), new TypeToken<Participant>() {
+        }.getType());
+
+        response.setResult(result);
+
+        response.setUniqueId(chatMessage.getUniqueId());
+
+        response.setCache(false);
+
+        return response;
+    }
+
+    public static String createEndRecordCall(StartOrEndCallRecordRequest request, String uniqueId) throws PodChatException  {
+
+        AsyncMessage message = new AsyncMessage();
+        message.setType(ChatMessageType.Constants.END_RECORDE_CALL);
+        message.setToken(CoreConfig.token);
+        message.setSubjectId(request.getSubjectId());
         message.setTokenIssuer(CoreConfig.tokenIssuer);
         message.setUniqueId(uniqueId);
         message.setTypeCode(Util.isNullOrEmpty(request.getTypeCode()) ? request.getTypeCode() : CoreConfig.typeCode);