+++ podchat/src/main/java/com/fanap/podchat/chat/Chat.java
@@ -50,6 +50,7 @@ import com.fanap.podchat.call.request_model.GetCallHistoryRequest;
 import com.fanap.podchat.call.request_model.GetCallParticipantsRequest;
 import com.fanap.podchat.call.request_model.MuteUnMuteCallParticipantRequest;
 import com.fanap.podchat.call.request_model.RejectCallRequest;
+import com.fanap.podchat.call.request_model.StartOrEndCallRecordRequest;
 import com.fanap.podchat.call.request_model.TerminateCallRequest;
 import com.fanap.podchat.call.result_model.CallCancelResult;
 import com.fanap.podchat.call.result_model.CallCreatedResult;
@@ -1149,6 +1150,12 @@ public class Chat extends AsyncAdapter {
             case Constants.REJECT_CALL:
                 handleOnCallRequestRejected(chatMessage);
                 break;
+            case Constants.START_RECORDE_CALL:
+                handleOnStartedCallRecord(chatMessage);
+                break;
+            case Constants.END_RECORDE_CALL:
+                handleOnEndedCallRecord(chatMessage);
+                break;
             case Constants.CANCEL_GROUP_CALL:
                 handleOnCallParticipantCanceledCall(chatMessage);
                 break;
@@ -2132,6 +2139,34 @@ public class Chat extends AsyncAdapter {
         ChatResponse<CallRequestResult> response
                 = CallAsyncRequestsManager.handleOnRejectCallRequest(chatMessage);
         listenerManager.callOnCallRequestRejected(response);
+
+    }
+
+    private void handleOnStartedCallRecord(ChatMessage chatMessage) {
+
+        if (sentryResponseLog) {
+            showLog("RECORD_CALL_STARTED", gson.toJson(chatMessage));
+        } else {
+            showLog("RECORD_CALL_STARTED");
+        }
+
+        ChatResponse<Participant> response
+                = CallAsyncRequestsManager.handleStartedRecordCallResponse(chatMessage);
+        listenerManager.callonCallRecordStarted(response);
+
+    }
+
+    private void handleOnEndedCallRecord(ChatMessage chatMessage) {
+
+        if (sentryResponseLog) {
+            showLog("RECORD_CALL_ENDED", gson.toJson(chatMessage));
+        } else {
+            showLog("RECORD_CALL_ENDED");
+        }
+        ChatResponse<Participant> response
+                = CallAsyncRequestsManager.handleStartedRecordCallResponse(chatMessage);
+
+        listenerManager.callonCallRecordEnded(response);
 
     }
 
@@ -3347,7 +3382,72 @@ public class Chat extends AsyncAdapter {
                 return uniqueId;
             }
 
-            sendAsyncMessage(message, AsyncAckType.Constants.WITHOUT_ACK, "SEND_STOP_BOT_REQUEST");
+            sendAsyncMessage(message, AsyncAckType.Constants.WITHOUT_ACK, "SEND_GET_USER_BOTS_REQUEST");
+
+        } else {
+            onChatNotReady(uniqueId);
+        }
+        return uniqueId;
+
+    }
+
+    /**
+     * @param request request to start call recording
+     *
+     *
+     *            */
+
+    public String startCallRecord(StartOrEndCallRecordRequest request) {
+
+        String uniqueId = generateUniqueId();
+
+        if (chatReady) {
+
+            String message = null;
+            try {
+                message = CallAsyncRequestsManager.createStartRecordCall(request, uniqueId);
+            } catch (PodChatException e) {
+                new PodThreadManager().doThisAndGo(() -> {
+                    e.setUniqueId(uniqueId);
+                    e.setToken(getToken());
+                    captureError(e);
+                });
+                return uniqueId;
+            }
+
+            sendAsyncMessage(message, AsyncAckType.Constants.WITHOUT_ACK, "SEND_START_CALL_RECORD_REQUEST");
+
+        } else {
+            onChatNotReady(uniqueId);
+        }
+        return uniqueId;
+
+    }
+    /**
+     * @param request request to end call recording
+     *
+     *
+     *            */
+
+    public String endCallRecord(StartOrEndCallRecordRequest request) {
+
+        String uniqueId = generateUniqueId();
+
+        if (chatReady) {
+
+            String message = null;
+            try {
+                message = CallAsyncRequestsManager.createEndRecordCall(request, uniqueId);
+            } catch (PodChatException e) {
+                new PodThreadManager().doThisAndGo(() -> {
+                    e.setUniqueId(uniqueId);
+                    e.setToken(getToken());
+                    captureError(e);
+                });
+                return uniqueId;
+            }
+
+            sendAsyncMessage(message, AsyncAckType.Constants.WITHOUT_ACK, "SEND_END_CALL_RECORD_REQUEST");
 
         } else {
             onChatNotReady(uniqueId);